+++ src/service/payment/movieTicket/factory.ts
@@ -5,6 +5,7 @@ import * as factory from '../../../factory';
 
 export function createSeatInfoSyncIn(params: {
     paymentMethodType: factory.paymentMethodType.MGTicket | factory.paymentMethodType.MovieTicket;
+    paymentMethodId: string;
     movieTickets: factory.chevre.paymentMethod.paymentCard.movieTicket.IMovieTicket[];
     event: factory.event.screeningEvent.IEvent;
     order: factory.order.ISimpleOrder;
@@ -61,7 +62,8 @@ export function createSeatInfoSyncIn(params: {
         kgygishCd: movieTicketPaymentAccepted.movieTicketInfo.kgygishCd,
         yykDvcTyp: mvtkapi.mvtk.services.seat.seatInfoSync.ReserveDeviceType.EntertainerSitePC, // 予約デバイス区分
         trkshFlg: mvtkapi.mvtk.services.seat.seatInfoSync.DeleteFlag.False, // 取消フラグ
-        kgygishSstmZskyykNo: params.order.orderNumber, // 興行会社システム座席予約番号
+        // kgygishSstmZskyykNo: params.order.orderNumber, // 興行会社システム座席予約番号
+        kgygishSstmZskyykNo: params.paymentMethodId, // 興行会社システム座席予約番号
         kgygishUsrZskyykNo: params.order.confirmationNumber.toString(), // 興行会社ユーザー座席予約番号
         jeiDt: moment(event.startDate)
             .tz('Asia/Tokyo')