+++ src/service/transaction/placeOrderInProgress/potentialActions/payMovieTicket.ts
@@ -1,5 +1,7 @@
 import * as factory from '../../../../factory';
 
+const USE_SEPARATE_MOVIE_TICKET_PAYMENT = process.env.USE_SEPARATE_MOVIE_TICKET_PAYMENT === '1';
+
 export async function createPayMovieTicketActions(params: {
     order: factory.order.IOrder;
     potentialActions?: factory.transaction.placeOrder.IPotentialActionsParams;
@@ -26,14 +28,14 @@ export async function createPayMovieTicketActions(params: {
             });
 
     if (authorizeMovieTicketActions.length > 0) {
-        payMovieTicketActions.push({
-            project: params.transaction.project,
-            typeOf: <factory.actionType.PayAction>factory.actionType.PayAction,
-            object: authorizeMovieTicketActions
-                .map((a) => {
-                    const result = <factory.action.authorize.paymentMethod.movieTicket.IResult>a.result;
+        if (USE_SEPARATE_MOVIE_TICKET_PAYMENT) {
+            authorizeMovieTicketActions.forEach((a) => {
+                const result = <factory.action.authorize.paymentMethod.movieTicket.IResult>a.result;
 
-                    return {
+                payMovieTicketActions.push({
+                    project: params.transaction.project,
+                    typeOf: <factory.actionType.PayAction>factory.actionType.PayAction,
+                    object: [{
                         typeOf: <factory.action.trade.pay.TypeOfObject>'PaymentMethod',
                         paymentMethod: {
                             accountId: result.accountId,
@@ -44,21 +46,56 @@ export async function createPayMovieTicketActions(params: {
                             typeOf: <factory.paymentMethodType.MovieTicket>result.paymentMethod
                         },
                         movieTickets: a.object.movieTickets
-                    };
-                }),
-            agent: params.transaction.agent,
-            purpose: {
-                project: params.order.project,
-                typeOf: params.order.typeOf,
-                seller: params.order.seller,
-                customer: params.order.customer,
-                confirmationNumber: params.order.confirmationNumber,
-                orderNumber: params.order.orderNumber,
-                price: params.order.price,
-                priceCurrency: params.order.priceCurrency,
-                orderDate: params.order.orderDate
-            }
-        });
+                    }],
+                    agent: params.transaction.agent,
+                    purpose: {
+                        project: params.order.project,
+                        typeOf: params.order.typeOf,
+                        seller: params.order.seller,
+                        customer: params.order.customer,
+                        confirmationNumber: params.order.confirmationNumber,
+                        orderNumber: params.order.orderNumber,
+                        price: params.order.price,
+                        priceCurrency: params.order.priceCurrency,
+                        orderDate: params.order.orderDate
+                    }
+                });
+            });
+        } else {
+            payMovieTicketActions.push({
+                project: params.transaction.project,
+                typeOf: <factory.actionType.PayAction>factory.actionType.PayAction,
+                object: authorizeMovieTicketActions
+                    .map((a) => {
+                        const result = <factory.action.authorize.paymentMethod.movieTicket.IResult>a.result;
+
+                        return {
+                            typeOf: <factory.action.trade.pay.TypeOfObject>'PaymentMethod',
+                            paymentMethod: {
+                                accountId: result.accountId,
+                                additionalProperty: (Array.isArray(result.additionalProperty)) ? result.additionalProperty : [],
+                                name: result.name,
+                                paymentMethodId: result.paymentMethodId,
+                                totalPaymentDue: result.totalPaymentDue,
+                                typeOf: <factory.paymentMethodType.MovieTicket>result.paymentMethod
+                            },
+                            movieTickets: a.object.movieTickets
+                        };
+                    }),
+                agent: params.transaction.agent,
+                purpose: {
+                    project: params.order.project,
+                    typeOf: params.order.typeOf,
+                    seller: params.order.seller,
+                    customer: params.order.customer,
+                    confirmationNumber: params.order.confirmationNumber,
+                    orderNumber: params.order.orderNumber,
+                    price: params.order.price,
+                    priceCurrency: params.order.priceCurrency,
+                    orderDate: params.order.orderDate
+                }
+            });
+        }
     }
 
     return payMovieTicketActions;